Add CSS demo (#511)
