+++ components/demos/index.jsx
@@ -2,7 +2,7 @@ export { GettingStartedDemo } from './GettingStarted';
 export { default as AccordionDemo } from './Accordion/stitches';
 export { default as AlertDialogDemo } from './AlertDialog/stitches';
 export { default as AspectRatioDemo } from './AspectRatio/stitches';
-export { default as AvatarDemo } from './Avatar';
+export { default as AvatarDemo } from './Avatar/stitches';
 export { default as CheckboxDemo } from './Checkbox';
 export { default as CollapsibleDemo } from './Collapsible';
 export { default as ContextMenuDemo } from './ContextMenu';
