+++ data/primitives/components/avatar/1.0.1.mdx
@@ -14,9 +14,7 @@ name: avatar
   <AvatarDemo />
 </HeroContainer>
 
-```jsx hero template=Avatar
-
-```
+<HeroCodeBlock folder="Avatar" />
 
 <Highlights
   features={[
